refactored to include AutoUpdateComponent in the statusBarItemsInjectable
