+++ src/common/application-update/progress-of-update-download/progress-of-update-download.injectable.ts
@@ -3,11 +3,10 @@
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
 import { getInjectable } from "@ogre-tools/injectable";
-import type { JsonObject } from "type-fest";
 import createSyncBoxInjectable from "../../utils/sync-box/create-sync-box.injectable";
 import { syncBoxInjectionToken } from "../../utils/sync-box/sync-box-injection-token";
 
-export interface ProgressOfDownload extends JsonObject {
+export interface ProgressOfDownload {
   percentage: number;
 }
